+++ react/typings/vtex.styleguide.d.ts
@@ -1,0 +1,14 @@
+declare module 'vtex.styleguide' {
+  import {ReactElement} from "react"
+
+  export const Badge: ReactElement
+  export const Button: ReactElement
+  export const Card: ReactElement
+  export const EmptyState: ReactElement
+  export const IconArrowBack: ReactElement
+  export const Input: ReactElement
+  export const Pagination: ReactElement
+  export const Spinner: ReactElement
+  export const Table: ReactElement
+  export const Toggle: ReactElement
+}